[Scan] Look for Reshift clusters without logging
This commit builds on the previous Redshift SSL scan functionality to
look for lack of logging. Both the SSL and logging scans involves
looking at parameter groups.

So, this commit refactors things into a `ParameterGroupScan` that can
be used by both the SSL and logging scans.

It also provides to eliminating flagging the default parameter group
if it isn't used. This is because the default group can't actually
be edited.