[Scan] Look for RDS instances with a single AZ
This commit adds support for finding RDS instances with
a single AZ. This could lead to denial of service in the
system.

Some refactoring was done to accomplish this. The `multiaz`
scan is essentially the same as the existing `encryption`
scan since they both just look for property name/value
states. A generic `PropertyScan` class was created that
can be inherited to match particular property states.

Example:

 * LOW * RDS instances without multiple availability zones (rds.multiaz)
{
    "rds_count": 1,
    "flagged_rds_count": 1,
    "instances": {
        "us-east-1": [
            "airflow"
        ]
    }
}