[scan] Implement scan type exclusions
This patch resolves issue #21 to implement scan type exclusions.
Previously, a list of scan suite names or scan names could be
specified on the CLI as *positional arguments*:

  $ scanamabob scan ec2 iam.mfa

However, in some cases you might want the ability to run
every suite except `ec2` or every scan in `iam` *except*
`iam.mfa`. This patches add that ability in the form of
a "scan set".

All positional arguments on the command line are treated
as scan suite or scan names that make up a scan set according
to the following rules:

1. No specified positional arguments means all scans are run.
2. Positional arguments may be optionally prefixed with a '+'
   or '-'.
3. Positional arguments with *no* prefix are included in
   the scan set.
4. Positional arguments preceded with a '+' are included in
   the scan set.
5. Any positional arguments that define inclusions define the
   default state of the scan set.  It is no longer the full
   set when inclusions are used.
6. Positional arguments preceded with a '-' are excluded from
   the scan set.

Because `argparse` does not allow positional arguments to have
a '-' prefix, a pre-pass is done that performs the following
mapping:

- `+foo` -> `--include=foo`
- `-foo` -> `--exclude=foo`

The `--include/exclude` options may be used directly. The `+/-`
form is simply shorter.

Examples:

  # Run everything.
  $ scanamabob scan

  # Run the IAM suite without the MFA scan.
  $ scanamabob scan +iam -iam.mfa

  # Run everything except EC2.
  scanamabob scan -ec2

  # Run nothing.
  $ scanamabob scan +iam -iam

  # Run just RDS.
  $ scanamabob scan rds