[RDS] Make the MultiAZ scan cluster aware
This patch fixes issue #37.  If a DB instance lives in a cluster, then
any properties defined on the cluster instance take precedence. For
example, the cluster `MultiAZ` property would take precedence over
the DB instance one.

This was causing false positive to be reported for DB instances in
any `MultiAZ` Aurora cluster.